requirements.txt
python-telegram-bot==22.3
huggingface-hub
Pillow
requests
python-dotenv
flask==2.3.2
gunicorn
aiohttp